Refactor node download logic
