+++ src/index.node.ts
@@ -1,16 +1,9 @@
 export { SkynetClient } from "./client/node";
 export { deriveChildSeed, genKeyPairAndSeed, genKeyPairFromSeed } from "./crypto";
-export {
-  MAX_REVISION,
-  defaultPortalUrl,
-  defaultSkynetPortalUrl,
-  getRelativeFilePath,
-  getRootDirectory,
-  parseSkylink,
-  uriHandshakePrefix,
-  uriHandshakeResolverPrefix,
-  uriSkynetPrefix,
-} from "./utils";
+export { parseSkylink, uriHandshakePrefix, uriHandshakeResolverPrefix, uriSkynetPrefix } from "./utils/skylink";
+export { MAX_REVISION } from "./utils/number";
+export { defaultPortalUrl, defaultSkynetPortalUrl } from "./utils/url";
+export { getRelativeFilePath, getRootDirectory } from "./utils/file";
 
 // Export types.
 
@@ -20,4 +13,4 @@ export type { CustomDownloadOptions, ResolveHnsResponse } from "./download";
 export type { CustomGetEntryOptions, CustomSetEntryOptions, SignedRegistryEntry, RegistryEntry } from "./registry";
 export type { CustomGetJSONOptions, CustomSetJSONOptions, VersionedEntryData } from "./skydb";
 export type { CustomUploadOptions, UploadRequestResponse } from "./upload/index";
-export type { ParseSkylinkOptions } from "./utils";
+export type { ParseSkylinkOptions } from "./utils/skylink";
